:art: Move navbar component to cube module and delete NavbarService
