+++ src/app/auth/email-verified/email-verified.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthService } from '../../core/auth.service';
-import { NavbarService } from 'app/core/navbar.service';
 
 @Component({
   selector: 'clark-email-verified',
