+++ src/app/auth/login/login.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthService } from '../../core/auth.service';
 import { Router, ActivatedRoute } from '@angular/router';
-import { NavbarService } from '../../core/navbar.service';
 
 @Component({
   selector: 'clark-login',
