+++ src/app/clark.component.html
@@ -1,4 +1,3 @@
-<clark-navbar *ngIf = "isSupportedBrowser"></clark-navbar>
 <router-outlet></router-outlet>
 <clark-popup *ngIf="isOldVersion" (closed)="isOldVersion = false">
   <div #popupInner class="popup-inner">
