+++ src/app/cube/cube.component.html
@@ -1,3 +1,4 @@
+<clark-navbar *ngIf = "true"></clark-navbar>
 <clark-toaster [content]="noteService.content"></clark-toaster>
 <app-contextmenu [content]="modalService.contextMenuContent"></app-contextmenu>
 <app-dialogmenu [content]="modalService.dialogMenuContent"></app-dialogmenu>
