+++ src/app/cube/details/error-status/error-status.component.ts
@@ -1,22 +1,18 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
+import { Component, Input } from '@angular/core';
+import { Router } from '@angular/router';
 import { AuthService } from 'app/core/auth.service';
-import { NavbarService } from 'app/core/navbar.service';
 
 @Component({
   selector: 'clark-error-status',
   templateUrl: './error-status.component.html',
   styleUrls: ['./error-status.component.scss']
 })
-export class ErrorStatusComponent implements OnInit {
+export class ErrorStatusComponent {
 
   @Input() statusCode: number;
   @Input() redirectUrl: string;
 
   constructor(private router: Router, private auth: AuthService) { }
-
-  ngOnInit() {
-  }
 
   login() {
     this.router.navigate(['auth'], { queryParams: { redirectUrl: this.redirectUrl } });
