+++ src/app/cube/shared/cube-shared.module.ts
@@ -17,6 +17,8 @@ import { ClickOutsideModule } from 'ng-click-outside';
 import { EscapeHtmlPipe } from '../../shared/pipes/keep-html.pipe';
 import { SharedModule } from '../../shared/shared.module';
 import { FeaturedComponent } from './featured/featured.component';
+import { NavbarComponent } from './navbar/navbar.component';
+import { MessageComponent } from './navbar/message/message.component';
 
 @NgModule({
   imports: [
@@ -28,11 +30,15 @@ import { FeaturedComponent } from './featured/featured.component';
     SharedModule
   ],
   exports: [
+    NavbarComponent,
+    MessageComponent,
     FooterComponent,
     EscapeHtmlPipe,
     FeaturedComponent
   ],
   declarations: [
+    NavbarComponent,
+    MessageComponent,
     FooterComponent,
     EscapeHtmlPipe,
     FeaturedComponent
