+++ src/app/cube/shared/navbar/navbar.component.html
@@ -1,0 +1,96 @@
+<clark-message></clark-message>
+<div class="topbar" [ngClass]="{'version': version, 'mobile': isMobile}">
+  <div class="inner flex h left-right">
+    <div class="logo flex h left-right" [ngClass]="{'noShow': searchDown && !isMobile}" routerLink="/home">
+      <div class="logo-image"></div>
+      <div id="clark-logo" class="logo-link">CLARK</div>
+      <div *ngIf="version" class="version-label">{{version}}</div>
+    </div>
+    <div class="clark-search-wrapper" [ngClass]="{'down': searchDown, 'mobile': isMobile, 'overflow': searchOverflow}">
+      <div class="clark-search-inner">
+        <clark-search [focus]="searchFocusSubject" [blur]="searchBlurSubject" (didFocus)="showSearch();" (didBlur)="hideSearch()"></clark-search>
+        <div class="close" id="search-close" (click)="hideSearch();"><i class="fal fa-times"></i></div>
+      </div>
+    </div>
+    <div id="clark-user" class="user flex h flex-end" [ngClass]="{'loggedin': loggedin, 'mobile': isMobile, 'noShow': searchDown && !isMobile}">
+      <ng-container *ngIf="isMobile; else userTemplate" [ngTemplateOutlet]="mobileUserTemplate"></ng-container>
+    </div>
+
+    <ng-template #userTemplate>
+      <ng-container *ngIf="loggedin">
+        <div class="contributor" id="contributor-link" #contributorDropdownButtonElement (click)="contributorDropdown(true)"><span>Contribute</span> <i class="fas fa-chevron-down"></i></div>
+        <div class="right library" id="library-link" routerLink="/library"></div>
+        <div id="clark-user-dropdown" #userDropdownButtonElement class="left flex h flex-end user-menu" (click)="userDropdown(true)">
+          <div class="navbar-gravatar" id="gravatar" [ngStyle]="{'background-image': 'url(' + gravatarImage(200) + ')'}"></div>
+          <i class="fas fa-chevron-down"></i>
+        </div>
+
+        <clark-context-menu *ngIf="contributorDown" [anchor]="contributorDropdownButtonElement" [offset]="{ top: 10, left: 5 }" (close)="contributorDropdown()">
+          <div #contextMenu>
+            <ul>
+              <li routerLink="/onion/dashboard" (click)="contributorDropdown()"><i class="far fa-list"></i>Your Dashboard</li>
+              <li routerLink="/onion/learning-object-builder" (click)="contributorDropdown()"><i class="far fa-plus"></i>Create a Learning Object</li>
+            </ul>
+          </div>
+        </clark-context-menu>
+        
+        <clark-context-menu *ngIf="userDown" [anchor]="userDropdownButtonElement" [offset]="{ top: 10, left: 5 }" (close)="userDropdown()">
+          <div #contextMenu id="context-popup">
+            <ul>
+              <li *ngIf="authService.hasCuratorAccess()" routerLink="/admin" (click)="userDropdown()"><i class="far fa-tools"></i> {{ authService.hasEditorAccess() ? 'Admin' : 'Curator' }} Dashboard</li>
+              <li [routerLink]="['users', authService.user?.username]" (click)="userDropdown()"><i class="far fa-user"></i>Your Profile</li>
+              <li routerLink="/auth/forgot-password" (click)="userDropdown()"><i class="far fa-unlock-alt"></i>Change Password</li>
+              <li (click)="logout(); userDropdown()"><i class="far fa-sign-out"></i>Sign Out</li>
+            </ul>
+          </div>
+        </clark-context-menu>
+      </ng-container>
+      <ng-container *ngIf="!loggedin">
+        <div class="login flex h flex-end" id="clark-sign-in" [routerLink]="['/auth/login']" [queryParams]="{redirectUrl: url}"><div>Sign in</div> <i class="fas fa-sign-in-alt"></i></div>
+        <div class="register" id="clark-register" routerLink="/auth/register"><div>Register</div> <i class="fas fa-user-plus"></i></div>
+      </ng-container>
+    </ng-template>
+
+    <ng-template #mobileUserTemplate>
+      <div class="mobile-search-option" (click)="showSearch()"><i class="far fa-search"></i></div>
+      <div class="hamburger"[ngClass]="{'active': menuOpen}" (click)="menuOpen = !menuOpen">
+        <div class="bar"></div>
+        <div class="bar"></div>
+        <div class="bar"></div>
+      </div>
+    </ng-template>
+  </div>
+</div>
+
+<!-- Overlay under mobile slideout -->
+<div class="slideout-underlay" [ngClass]="{'active': menuOpen || searchDown, 'underNav': !isMobile || searchDown}" (click)="menuOpen = false; hideSearch();"></div>
+
+<!-- Mobile menu slideout -->
+<div class="slideout" *ngIf="isMobile" [ngClass]="{'active': menuOpen}">
+  <div class="slideout-section user" routerLink="{{ loggedin ? '/users/' + authService.username : '/auth/login' }}">
+    <ng-container *ngIf="loggedin; else noUserGreeting">
+      <div class="img" [ngStyle]="{'background-image': 'url(' + gravatarImage(200) + ')'}"></div>
+      <div class="right">
+        <div class="greeting">Howdy, {{ authService.firstName | titlecase }}!</div>
+        <a>View your profile</a>
+      </div>
+    </ng-container>
+    <ng-template #noUserGreeting>
+      <div class="greeting">Hi there!<br />We're happy you're here!</div>
+    </ng-template>
+  </div>
+  <div class="slideout-section menu-items">
+    <ul>
+      <li *ngIf="loggedin" routerLink="/onion/dashboard">Your Dashboard</li>
+      <li *ngIf="loggedin" routerLink="/library">Your Library</li>
+      <li *ngIf="loggedin" (click)="logout()">Logout</li>
+    </ul>
+  </div>
+  <div class="slideout-section auth" *ngIf="!loggedin">
+    <div class="btn-group vertical">
+      <button class="button good" routerLink="/auth/register">Register</button>
+      <button class="button neutral" routerLink="/auth/login">Log In</button>
+    </div>
+  </div>
+</div>
+
